Move shouldStart()
